changes and structure of folders
