--- src/pages/checkout-v2/useCheckoutV2.js
@@ -1,5 +1,0 @@
-import useCheckout from '../checkout/useCheckout';
-
-const useCheckoutV2 = () => useCheckout();
-
-export default useCheckoutV2;
